refactor: 💡 simplified root route and add fallback routing
