+++ src/App.tsx
@@ -1,17 +1,31 @@
-import { useEffect } from 'react';
-import { useLocation, useNavigate } from 'react-router-dom';
-import Main from './containers/Main';
+import { Outlet } from 'react-router-dom';
+import styled from 'styled-components';
+import TabsNav from './components/TabsNav';
 
+export const Wrapper = styled.div`
+  width: 100vw;
+  height: 100vh;
+  background-color: #fff;
+`;
+
+export const TabView = styled.div`
+  display: flex;
+  flex-direction: column;
+`;
+
+/**
+ * Renders the main application component.
+ * It handles the navigation and renders the tab bar and tab view.
+ */
 function App() {
-  const path = useLocation().pathname;
-  const navigate = useNavigate();
-
-  useEffect(() => {
-    if (path === '/') {
-      navigate('/search');
-    }
-  }, [path, navigate]);
-  return <Main />;
+  return (
+    <Wrapper>
+      {/* tab bar */}
+      <TabsNav />
+      {/* tab view */}
+      <Outlet />
+    </Wrapper>
+  );
 }
 
 export default App;
