+++ src/components/TabsNav/Styled.ts
