+++ src/components/TabsNav/TabsNav.tsx
