--- src/containers/Main/Styled.ts
@@ -1,11 +1,0 @@
-import styled from 'styled-components';
-
-export const Wrapper = styled.div`
-  width: 100vw;
-  height: 100vh;
-  background-color: #fff;
-`;
-export const TabView = styled.div`
-  display: flex;
-  flex-direction: column;
-`;
